Rework easom_tl
